requirements.txt
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
scikit-learn>=1.3.0
xgboost>=1.7.0
shap>=0.42.0
matplotlib>=3.7.0
joblib>=1.3.0
cmcrameri>=1.7
pyyaml>=6.0
tqdm>=4.66.0
